WIP - updating libs so that there is more flexibily in mongo ops
