+++ config.yaml
@@ -1,0 +1,10 @@
+# Copyright 2024 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+options:
+  auto-delete:
+    type: boolean
+    description: |
+      When a relation is removed, auto-delete ensures that any relevant databases
+      associated with the relation are also removed
+    default: false
