+++ lib/charms/mongodb/v1/mongodb.py
@@ -6,12 +6,12 @@
 import logging
 import re
 from dataclasses import dataclass
-from typing import Dict, List, Optional, Set
+from typing import Dict, Optional, Set
 from urllib.parse import quote_plus
 
+from charms.mongodb.v0.mongo import MongoConfiguration, MongoConnection, NotReadyError
 from bson.json_util import dumps
-from pymongo import MongoClient
-from pymongo.errors import OperationFailure, PyMongoError
+from pymongo.errors import OperationFailure
 from tenacity import (
     RetryError,
     Retrying,
@@ -43,7 +43,7 @@ class FailedToMovePrimaryError(Exception):
 
 
 @dataclass
-class MongoDBConfiguration:
+class MongoDBConfiguration(MongoConfiguration):
     """Class for MongoDB configuration.
 
     — replset: name of replica set, needed for connection URI.
@@ -90,11 +90,7 @@ class MongoDBConfiguration:
         )
 
 
-class NotReadyError(PyMongoError):
-    """Raised when not all replica set members healthy or finished initial sync."""
-
-
-class MongoDBConnection:
+class MongoDBConnection(MongoConnection):
     """In this class we create connection object to MongoDB.
 
     Real connection is created on the first call to MongoDB.
@@ -124,49 +120,7 @@ class MongoDBConnection:
             direct: force a direct connection to a specific host, avoiding
                     reading replica set configuration and reconnection.
         """
-        self.mongodb_config = config
-
-        if uri is None:
-            uri = config.uri
-
-        self.client = MongoClient(
-            uri,
-            directConnection=direct,
-            connect=False,
-            serverSelectionTimeoutMS=1000,
-            connectTimeoutMS=2000,
-        )
-        return
-
-    def __enter__(self):
-        """Return a reference to the new connection."""
-        return self
-
-    def __exit__(self, object_type, value, traceback):
-        """Disconnect from MongoDB client."""
-        self.client.close()
-        self.client = None
-
-    @property
-    def is_ready(self) -> bool:
-        """Is the MongoDB server ready for services requests.
-
-        Returns:
-            True if services is ready False otherwise. Retries over a period of 60 seconds times to
-            allow server time to start up.
-
-        Raises:
-            ConfigurationError, ConfigurationError, OperationFailure
-        """
-        try:
-            for attempt in Retrying(stop=stop_after_delay(60), wait=wait_fixed(3)):
-                with attempt:
-                    # The ping command is cheap and does not require auth.
-                    self.client.admin.command("ping")
-        except RetryError:
-            return False
-
-        return True
+        MongoConnection.__init__(self, config, uri, direct)
 
     @retry(
         stop=stop_after_attempt(3),
@@ -347,107 +301,6 @@ class MongoDBConnection:
 
         logger.debug("rs_config: %r", rs_config)
         self.client.admin.command("replSetReconfig", rs_config)
-
-    def create_user(self, config: MongoDBConfiguration):
-        """Create user.
-
-        Grant read and write privileges for specified database.
-        """
-        self.client.admin.command(
-            "createUser",
-            config.username,
-            pwd=config.password,
-            roles=self._get_roles(config),
-            mechanisms=["SCRAM-SHA-256"],
-        )
-
-    def update_user(self, config: MongoDBConfiguration):
-        """Update grants on database."""
-        self.client.admin.command(
-            "updateUser",
-            config.username,
-            roles=self._get_roles(config),
-        )
-
-    def set_user_password(self, username, password: str):
-        """Update the password."""
-        self.client.admin.command(
-            "updateUser",
-            username,
-            pwd=password,
-        )
-
-    def create_role(self, role_name: str, privileges: dict, roles: dict = []):
-        """Creates a new role.
-
-        Args:
-            role_name: name of the role to be added.
-            privileges: privileges to be associated with the role.
-            roles: List of roles from which this role inherits privileges.
-        """
-        try:
-            self.client.admin.command(
-                "createRole", role_name, privileges=[privileges], roles=roles
-            )
-        except OperationFailure as e:
-            if not e.code == 51002:  # Role already exists
-                logger.error("Cannot add role. error=%r", e)
-                raise e
-
-    @staticmethod
-    def _get_roles(config: MongoDBConfiguration) -> List[dict]:
-        """Generate roles List."""
-        supported_roles = {
-            "admin": [
-                {"role": "userAdminAnyDatabase", "db": "admin"},
-                {"role": "readWriteAnyDatabase", "db": "admin"},
-                {"role": "userAdmin", "db": "admin"},
-            ],
-            "monitor": [
-                {"role": "explainRole", "db": "admin"},
-                {"role": "clusterMonitor", "db": "admin"},
-                {"role": "read", "db": "local"},
-            ],
-            "backup": [
-                {"db": "admin", "role": "readWrite", "collection": ""},
-                {"db": "admin", "role": "backup"},
-                {"db": "admin", "role": "clusterMonitor"},
-                {"db": "admin", "role": "restore"},
-                {"db": "admin", "role": "pbmAnyAction"},
-            ],
-            "default": [
-                {"role": "readWrite", "db": config.database},
-            ],
-        }
-        return [role_dict for role in config.roles for role_dict in supported_roles[role]]
-
-    def drop_user(self, username: str):
-        """Drop user."""
-        self.client.admin.command("dropUser", username)
-
-    def get_users(self) -> Set[str]:
-        """Add a new member to replica set config inside MongoDB."""
-        users_info = self.client.admin.command("usersInfo")
-        return set(
-            [
-                user_obj["user"]
-                for user_obj in users_info["users"]
-                if re.match(r"^relation-\d+$", user_obj["user"])
-            ]
-        )
-
-    def get_databases(self) -> Set[str]:
-        """Return list of all non-default databases."""
-        system_dbs = ("admin", "local", "config")
-        databases = self.client.list_database_names()
-        return set([db for db in databases if db not in system_dbs])
-
-    def drop_database(self, database: str):
-        """Drop a non-default database."""
-        system_dbs = ("admin", "local", "config")
-        if database in system_dbs:
-            return
-        self.client.drop_database(database)
 
     def _is_primary(self, rs_status: Dict, hostname: str) -> bool:
         """Returns True if passed host is the replica set primary.
